% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{acc_matrix}
\alias{acc_matrix}
\title{Viability Classifier Accuracies}
\format{A matrix with 15 rows and columns
\describe{
  \item{rows}{Viability classifiers trained on each organoid line}
  \item{columns}{Validation data for each organoid line}
}}
\usage{
acc_matrix
}
\description{
The full accuracy matrix of classifiers for individual organoid lines (rows)
applied to features of all other lines (columns)
}
\keyword{datasets}
