% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{acc_matrix_reduced}
\alias{acc_matrix_reduced}
\title{Viability Classifier Accuracies for Two-Channel Classifiers}
\format{A list with 4 entries. Each entry is a 15 x 15 matrix.}
\usage{
acc_matrix_reduced
}
\description{
A list of the full accuracy matrices of classifiers for individual organoid
lines (rows) applied to features of all other lines (columns). The list has
four entries, one for each reduced classifier: only actin, only cell nuclei,
only dead cells, or actin + cell nuclei.
}
\keyword{datasets}
