% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{auc_ctg}
\alias{auc_ctg}
\title{CellTiter-Glo Data}
\format{Data frame with 12288 rows and 29 columns
\describe{
  \item{screen}{}
  \item{date}{}
  \item{operator}{}
  \item{mithras}{}
  \item{full_barcode}{}
  \item{Line}{}
  \item{plate}{}
  \item{library}{}
  \item{tag1}{}
  \item{tag2}{}
  \item{well}{}
  \item{row}{}
  \item{col}{}
  \item{pcount}{}
  \item{batch}{}
  \item{drug}{}
  \item{substance}{}
  \item{Concentration}{}
  \item{rack.position}{}
  \item{volume}{}
  \item{norm_fac}{}
  \item{pcount_norm}{}
  \item{med_ctrl}{}
  \item{viability}{}
  \item{rep}{}
  \item{s_AUC_fit}{}
  \item{s_AUC_actual}{}
  \item{l_AUC_fit}{}
  \item{l_AUC_actual}{}
}}
\usage{
auc_ctg
}
\description{
CellTiter-Glo viability scores and corresponding dose-response AUCs
}
\keyword{datasets}
