% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{auc_img_twoChannel}
\alias{auc_img_twoChannel}
\title{Dose-Response Data for Viability Classifier}
\format{Data frame with 10464 rows and 16 columns
\describe{
  \item{Product.Name}{}
  \item{Concentration}{}
  \item{Plate.ID}{}
  \item{Well.ID}{}
  \item{Num.Objects}{}
  \item{Percent.Dead}{Percent of organoids that are dead in the well}
  \item{Percent.Live}{Percent of organoids that are alive in the well}
  \item{Mean.Certainty.Live}{The mean classification accuracy of all organoids classified as alive}
  \item{Mean.Certainty.Dead}{The mean classification accuracy of all organoids classified as dead}
  \item{Line}{}
  \item{Replicate}{}
  \item{Layout}{}
  \item{s_AUC_fit}{}
  \item{s_AUC_actual}{}
  \item{l_AUC_fit}{}
  \item{l_AUC_actual}{}
}}
\usage{
auc_img_twoChannel
}
\description{
Dose-Response AUCs for Clinical Cancer Panel subset of image-based
viabilities as determined by the two-channel (actin + cell nuclei)
classifier.
}
\keyword{datasets}
