% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{drug_annotations}
\alias{drug_annotations}
\title{Drug Target Annotations}
\format{Data frame with 528 rows and 3 columns
\describe{
  \item{Drug}{}
  \item{Pathways}{}
  \item{Targets}{}
}}
\usage{
drug_annotations
}
\description{
Pathways and targets for all drugs in the screen.
}
\keyword{datasets}
