% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{drug_effect_metadata}
\alias{drug_effect_metadata}
\title{Drug-induced phenotype profile metadata}
\format{Data frame with 11620 rows and 6 columns
\describe{
  \item{AUC_Mean}{The mean SVM AUROC, i.e. classification performance,
  averaged over 10 cross validation iterations.}
  \item{AUC_Std}{The standard deviation of the SVM AUROC over 10 cross
  validation iterations.}
  \item{Distance}{The euclidean distance between the median features of
  negative controls and treated drug (after PCA-transformation).}
  \item{Line}{}
  \item{Drug}{}
  \item{Concentration}{}
}}
\usage{
drug_effect_metadata
}
\description{
Metadata belonging to 'drug_effect_profiles'. Row X in the metadata
corresponds to row X in the profiles (rownames are identical)
}
\keyword{datasets}
