% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{drug_effect_profiles}
\alias{drug_effect_profiles}
\title{Drug-induced phenotype profiles}
\format{Matrix with 11620 rows and 25 columns}
\usage{
drug_effect_profiles
}
\description{
Drug-induced phenotype vector as calculated by the SVMs. Each row in the
matrix corresponds to one treatment. A treatment consists of a line, drug,
and concentration (when applicable). The treatment ID can be read from the
row names with the format 'LINE.DRUG__CONCENTRATION' (for Clinical Cancer
Panel drugs) or 'LINE.DRUG' (for KiStem drugs). 'drug_effect_metadata' also
has this information in a more easily manageable table format.
}
\details{
Profiles are averaged over 10 cross validation iterations.
}
\keyword{datasets}
