% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_rda_files.R
\name{generate_rda_files_drugeffects}
\alias{generate_rda_files_drugeffects}
\title{Generate RDA files for Drug Effects}
\usage{
generate_rda_files_drugeffects()
}
\value{
A boolean indicating that the function ran without error.
}
\description{
Transforms the data generated by the Python code into RDA
    files used by the drug effects vignette. This function is not
    meant to be run and serves only as a 'log' for the conversion steps.

    The function assumes the current working directory is the base
    directory of the package, i.e. where the folders 'data', 'R', etc.
    are located.
}
\examples{
print(generate_rda_files_drugeffects)
}
\author{
Jan Sauer
}
