% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_enrichment.R
\name{get_cluster_enrichment}
\alias{get_cluster_enrichment}
\title{Hierarchical Cluster Enrichment Analysis}
\usage{
get_cluster_enrichment()
}
\arguments{
\item{clustering}{An hclust object.}

\item{labels}{A list of labels, each entry being a vector of integers
indicating the IDs of the group members. The IDs should match
'clustering$labels'}

\item{min_cluster_size}{The minimum cluster size to consider.}

\item{max_cluster_size}{THe maximum cluster size to consider. Defaults to
1/4 of the total number of elements or the size of the largest label
 group, depending on what is larger.}

\item{min_odds_ratio}{A lower limit for the odds ratio for a significant
cluster. This can be set to 0 if no limitation is desired.}

\item{max_annotated_labels}{Optionally filter the fisher test results and
use only the N most significant labels when generating the annotation
vector.}

\item{keep_only_best_cluster}{A boolean that determines if only the most
significant cluster for each pathway is kept. Not implemented yet.}
}
\value{
A list with two elements: a vector of significant annotations,
       corresponding to the clustered object IDs and the results of the
       fisher test
}
\description{
Find label enrichments in a hierarchical cluster object. This
    function is based on:
    Freudenberg et al., "CLEAN: CLustering Enrichment ANalysis",
    BMC Bioinformatics, 2009, 10:234
}
\details{
These functions perform an enrichment analysis for hierarchical clustering
based on pre-defined labels. The idea of this comes from
}
\examples{
print(get_cluster_enrichment)
}
\author{
Jan Sauer
}
