% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetPathwaysTargets.R
\name{get_mode_of_action}
\alias{get_mode_of_action}
\title{Get Mode of Action / Pathway}
\usage{
get_mode_of_action()
}
\value{
A list of pathways
}
\description{
Loads the annotation data and retrieves the pathway on which 
    drugs act.
}
\examples{
print(get_mode_of_action)
}
\author{
Jan Sauer
}
