% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lines.R
\name{get_replicate_ids_for_plates}
\alias{get_replicate_ids_for_plates}
\title{Get Replicate IDs for Plates}
\usage{
get_replicate_ids_for_plates()
}
\value{
A named vector containing the replicate IDs
}
\description{
Get hard-coded replicate IDs for all the plates.
}
\examples{
print(get_replicate_ids_for_plates)
}
\author{
Jan Sauer
}
