% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetPathwaysTargets.R
\name{get_targets}
\alias{get_targets}
\title{Get Drug Targets}
\usage{
get_targets()
}
\value{
A list of drug targets
}
\description{
Loads the annotation data and retrieves the targets on which 
    drugs act. There is some preprocessing involved here as a drug can 
    have multiple targets. In this case, a comma-separated list is returned.
}
\examples{
print(get_targets)
}
\author{
Jan Sauer
}
