% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mortality}
\alias{mortality}
\title{Viabilities of wells}
\format{Data frame with 32866 rows and 12 columns
\describe{
  \item{Product.Name}{}
  \item{Concentration}{}
  \item{Plate.ID}{}
  \item{Well.ID}{}
  \item{Num.Objects}{}
  \item{Percent.Dead}{Percent of organoids that are dead in the well}
  \item{Percent.Live}{Percent of organoids that are alive in the well}
  \item{Mean.Certainty.Live}{The mean classification accuracy of all organoids classified as alive}
  \item{Mean.Certainty.Dead}{The mean classification accuracy of all organoids classified as dead}
  \item{Line}{}
  \item{Replicate}{}
  \item{Layout}{}
}}
\usage{
mortality
}
\description{
Classification results for wells.
}
\keyword{datasets}
