% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mortality_reduced}
\alias{mortality_reduced}
\title{Viabilities of wells}
\format{A list with 4 entries. Each entry is a data frame with 32866 rows
and 12 columns.
\describe{
  \item{Product.Name}{}
  \item{Concentration}{}
  \item{Plate.ID}{}
  \item{Well.ID}{}
  \item{Num.Objects}{}
  \item{Percent.Dead}{Percent of organoids that are dead in the well}
  \item{Percent.Live}{Percent of organoids that are alive in the well}
  \item{Mean.Certainty.Live}{The mean classification accuracy of all organoids classified as alive}
  \item{Mean.Certainty.Dead}{The mean classification accuracy of all organoids classified as dead}
  \item{Line}{}
  \item{Replicate}{}
  \item{Layout}{}
}}
\usage{
mortality_reduced
}
\description{
A list of classification results for wells for all reduced-channel
classifiers. The list has four entries, one for each reduced classifier:
only actin, only cell nuclei, only dead cells, or actin + cell nuclei.
}
\keyword{datasets}
