% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{roc_aucs}
\alias{roc_aucs}
\title{ROC curve AUCs}
\format{Data frame with 3 columns
\describe{
  \item{clf_line}{The source line of the classifier being tested}
  \item{data_line}{The source line for the data being tested on}
  \item{AUC}{Area under receiver operating curve}
}}
\usage{
roc_aucs
}
\description{
AUC values for ROC curves found in 'roc_data'
}
\keyword{datasets}
