% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{roc_data}
\alias{roc_data}
\title{ROC curves for viability classifiers}
\format{Data frame with 5 columns
\describe{
  \item{Threshold}{The classification threshold being tested. Sometimes, a
  threshold of 2.0 is selected by python. This is only a technical measure
  to ensure that the ROC curve goes from (0, 0) to (1, 1).}
  \item{FalsePosRate}{False positive rate at given threshold}
  \item{TruePosRate}{True positive rate at given threshold}
  \item{clf_line}{The source line of the classifier being tested}
  \item{data_line}{The source line for the data being tested on}
}}
\usage{
roc_data
}
\description{
ROC curves for viability classifiers for all combinations of classifiers
applied to data of other lines.
}
\keyword{datasets}
