% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{well_features}
\alias{well_features}
\title{Well-averaged features}
\format{Matrix with 32939 rows and 3143 columns}
\usage{
well_features
}
\description{
Well-averages of processed single-organoid features. Note that this still
contains all features, even those that were too constant, i.e. resulted
in a value of NA/NaN after preprocessing. Row names indicate plate and
well ID. Note that some wells are missing, either because data was corrupted
(D013T01) or because imaging quality was too low for correct segmentation
(D010T01, D046T01, D054T01, and D055T01). Also note that D054T01 was only
treated with the clinical cancer panel library (L08).
}
\keyword{datasets}
